misc: refactor apis into a diff file
